requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
grpcio==1.64.0
grpcio-tools==1.64.0
openai==1.35.0
httpx<0.28
chromadb==0.5.0
numpy<2.0
sentence-transformers==3.0.0
prometheus-fastapi-instrumentator==7.0.0
prometheus-client==0.20.0
locust==2.29.0
python-dotenv==1.0.1
